+++ Save-MAML.ps1
@@ -91,7 +91,11 @@
             $theModuleHelpFile = Join-Path $theModuleCultureDir "$m-Help.xml" # Construct the path to the module help file (e.g. en-us\Module-Help.xml)
 
             & $getMAML @splat | # Convert the module help to MAML,
-                Set-Content -Encoding UTF8 -Path $theModuleHelpFile -Passthru: $PassThru # and write the file.
+                Set-Content -Encoding UTF8 -Path $theModuleHelpFile # and write the file.
+            
+            if ($Passthru) {
+                Get-Item -Path $theModuleHelpFile
+            }
          }
 
         if ($t -gt 1) {